+++ hooks/src/components/ImageGallery.tsx
@@ -1,26 +1,9 @@
-import React, { useState, useEffect } from "react";
-
-const API_KEY = import.meta.env.VITE_PIXABAY_API_KEY;
-
-interface Image {
-  id: number;
-  webformatURL: string;
-  tags: string;
-}
+import React, { useState } from 'react';
+import useImageSearch from '../hooks/useImageSearch';
 
 const ImageGallery: React.FC = () => {
-  const [images, setImages] = useState<Image[]>([]);
-  const [keyword, setKeyword] = useState("");
-
-  // useEffect를 사용하여 keyword가 변경될 때마다 이미지를 가져오는 로직을 작성해보세요.
-  useEffect(() => {
-    // 여기에 코드를 작성하세요.
-    // 힌트:
-    // 1. fetchImages 함수를 정의하세요.
-    // 2. fetchImages 함수 내부에서 Pixabay API를 호출하여 이미지를 가져오세요.
-    // 3. 가져온 이미지를 상태(images)에 업데이트하세요.
-    // 4. keyword가 존재할 때만 fetchImages 함수를 호출하세요.
-  }, [keyword]);
+  const [keyword, setKeyword] = useState('');
+  const { images } = useImageSearch(keyword);
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -34,14 +17,14 @@ const ImageGallery: React.FC = () => {
       </header>
       <section>
         <form onSubmit={handleSubmit}>
-          <label htmlFor="searchKeyword">검색:</label>
+          <label htmlFor='searchKeyword'>검색:</label>
           <input
-            type="text"
-            id="searchKeyword"
-            name="keyword"
-            placeholder="키워드 입력"
+            type='text'
+            id='searchKeyword'
+            name='keyword'
+            placeholder='키워드 입력'
           />
-          <button type="submit">Search</button>
+          <button type='submit'>Search</button>
         </form>
       </section>
       <section>
